♻️ Refactor: profile section 컴포넌트 분리
